+++ crates/threshold-signature-server/src/validator/tests.rs
@@ -23,7 +23,7 @@ use crate::{
         },
     },
     validator::{
-        api::{prune_old_holders, validate_new_reshare},
+        api::{is_signer_or_delete_parent_key, prune_old_holders, validate_new_reshare},
         errors::ValidatorErr,
     },
 };
@@ -54,7 +54,10 @@ async fn test_reshare() {
     let alice = AccountKeyring::AliceStash;
 
     let cxt = test_node_process_testing_state(true).await;
-    let (_validator_ips, _validator_ids) = spawn_testing_validators(true).await;
+
+    let add_parent_key_to_kvdb = true;
+    let (_validator_ips, _validator_ids) = spawn_testing_validators(add_parent_key_to_kvdb).await;
+
     let validator_ports = vec![3001, 3002, 3003];
     let api = get_api(&cxt.ws_url).await.unwrap();
     let rpc = get_rpc(&cxt.ws_url).await.unwrap();
@@ -221,4 +224,24 @@ async fn test_forbidden_keys() {
 
     let should_pass = check_forbidden_key("test");
     assert_eq!(should_pass.unwrap(), ());
+}
+
+#[tokio::test]
+#[serial]
+async fn test_deletes_key() {
+    initialize_test_logger().await;
+    clean_tests();
+
+    let dave = AccountKeyring::Dave;
+    let kv = setup_client().await;
+    let reservation = kv.kv().reserve_key(hex::encode(NETWORK_PARENT_KEY)).await.unwrap();
+    kv.kv().put(reservation, vec![10]).await.unwrap();
+
+    let is_proper_signer_result =
+        is_signer_or_delete_parent_key(&dave.to_account_id().into(), vec![], &kv).await.unwrap();
+    assert!(!is_proper_signer_result);
+
+    let has_key = kv.kv().exists(&hex::encode(NETWORK_PARENT_KEY)).await.unwrap();
+    assert!(!has_key);
+    clean_tests();
 }